+++ src/audio_effects.f90
@@ -1,3 +1,8 @@
+! Forsynth: a multitracks stereo sound synthesis project
+! License GPL-3.0-or-later
+! Vincent Magnin
+! Last modifications: 2023-03-28
+
 module audio_effects
     ! Various audio effects