+++ src/envelopes.f90
@@ -1,3 +1,8 @@
+! Forsynth: a multitracks stereo sound synthesis project
+! License GPL-3.0-or-later
+! Vincent Magnin
+! Last modifications: 2023-03-28
+
 module envelopes
     ! Subroutines generating envelopes
     ! https://en.wikipedia.org/wiki/Envelope_(music)