+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Added some scales in `demo3()`.
 
 ### Changed
-- Now under GNU GPLv3 license.
+- Now under GPL-3.0-or-later license.
 
 ### Fixed
 - A bug in add_karplus_strong().